Separate evaluation from solving
This changeset introduces a new function, ``session.solve``, which
always returns a list of Variable instances (possibly resolved ones,
that is Variable instances with no dependencies). In contrast,
``sessions.evaluate`` now returns non-variable values or warns for the
deprecation of the support for evaluating unresolved variables.

A couple of code factorization is also including in this changeset.